+++ vscode-ext/src/extension.ts
@@ -15,7 +15,12 @@ const versionCheckEndpoint = 'https://version.mirrord.dev/get-latest-version';
 
 let buttons: { toggle: vscode.StatusBarItem, settings: vscode.StatusBarItem };
 let globalContext: vscode.ExtensionContext;
-let k8sApi: CoreV1Api;
+
+function getK8sApi(): CoreV1Api {
+	let k8sConfig = new k8s.KubeConfig();
+	k8sConfig.loadFromDefault();
+	return k8sConfig.makeApiClient(k8s.CoreV1Api);
+}
 
 async function changeSettings() {
 	let agentNamespace = globalContext.workspaceState.get<string>('agentNamespace', 'default');
@@ -37,7 +42,7 @@ async function changeSettings() {
 			const namespaces: {
 				response: any;
 				body: V1NamespaceList;
-			} = await k8sApi.listNamespace();
+			} = await getK8sApi().listNamespace();
 			const namespaceNames = namespaces.body.items.map(namespace => namespace.metadata!.name!);
 			vscode.window.showQuickPick(namespaceNames, { placeHolder: 'Select namespace' }).then(async namespaceName => {
 				if (namespaceName === undefined) {
@@ -94,9 +99,6 @@ export async function activate(context: vscode.ExtensionContext) {
 	// TODO: Download mirrord according to platform
 	checkVersion(context.extension.packageJSON.version);
 	globalContext = context;
-	let k8sConfig = new k8s.KubeConfig();
-	k8sConfig.loadFromDefault();
-	k8sApi = k8sConfig.makeApiClient(k8s.CoreV1Api);
 
 	context.globalState.update('enabled', false);
 	vscode.debug.registerDebugConfigurationProvider('*', new ConfigurationProvider(), 2);
@@ -136,6 +138,7 @@ class ConfigurationProvider implements vscode.DebugConfigurationProvider {
 		}
 
 		const podNamespace = globalContext.workspaceState.get<string>('impersonatedPodNamespace', 'default');
+		let k8sApi = getK8sApi();
 		// Get pods from kubectl and let user select one to mirror
 		let pods: { response: any, body: V1PodList } = await k8sApi.listNamespacedPod(podNamespace);
 		let podNames = pods.body.items.map((pod) => pod.metadata!.name!);